Determine HTTPs redirect based on config presence
SSL certificates may have very restricted permissions and Nginx, running
as www or www-data during request time, may not be able to check if the
files exist. So checking for the config presence instead works.

'During request time' - because we're checking for \$server_name,
instead of a hardcoded hostname. TODO: check on startup, improve
performance.